--- sim/common/sod/items_sets/phase_1.go
@@ -1,78 +1,0 @@
-package item_sets
-
-import (
-	"github.com/wowsims/classic/sim/core"
-	"github.com/wowsims/classic/sim/core/stats"
-)
-
-///////////////////////////////////////////////////////////////////////////
-//                                 Cloth
-///////////////////////////////////////////////////////////////////////////
-
-var ItemSetBlackfathomInvokerVestaments = core.NewItemSet(core.ItemSet{
-	Name: "Twilight Invoker's Vestments",
-	Bonuses: map[int32]core.ApplyEffect{
-		2: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.SpellPower, 9)
-		},
-		3: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.MeleeHit, 1)
-			c.AddStat(stats.SpellHit, 1)
-		},
-	},
-})
-
-///////////////////////////////////////////////////////////////////////////
-//                                 Leather
-///////////////////////////////////////////////////////////////////////////
-
-var ItemSetBlackfathomSlayerLeather = core.NewItemSet(core.ItemSet{
-	Name: "Blackfathom Slayer's Leather",
-	Bonuses: map[int32]core.ApplyEffect{
-		2: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.AttackPower, 12)
-		},
-		3: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.MeleeHit, 1)
-			c.AddStat(stats.SpellHit, 1)
-		},
-	},
-})
-
-var ItemSetBlackfathomElementalistHide = core.NewItemSet(core.ItemSet{
-	Name: "Blackfathom Elementalist's Hide",
-	Bonuses: map[int32]core.ApplyEffect{
-		2: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.SpellPower, 12)
-		},
-		3: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.MeleeHit, 1)
-			c.AddStat(stats.SpellHit, 1)
-		},
-	},
-})
-
-///////////////////////////////////////////////////////////////////////////
-//                                 Mail
-///////////////////////////////////////////////////////////////////////////
-
-var ItemSetBlackfathomAvengerMail = core.NewItemSet(core.ItemSet{
-	Name: "Blackfathom Avenger's Mail",
-	Bonuses: map[int32]core.ApplyEffect{
-		2: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.AttackPower, 12)
-		},
-		3: func(agent core.Agent) {
-			c := agent.GetCharacter()
-			c.AddStat(stats.MeleeHit, 1)
-			c.AddStat(stats.SpellHit, 1)
-		},
-	},
-})
